+++ .prettierignore
@@ -1,3 +1,4 @@
 .next
 .layer0
 node_modules
+.swc
