+++ lib/helper.js
@@ -1,5 +1,25 @@
 export const relativizeURL = (str) => str.replace('https://layer0-docs-layer0-ecommmerce-api-example-default.layer0-limelight.link/', '/l0-api/')
 
+export const getOrigin = (req) => {
+  let origin
+  if (typeof window !== 'undefined') {
+    origin = window.location.origin
+  }
+  if (req) {
+    let hostURL = req.headers['host']
+    if (hostURL) {
+      hostURL = hostURL.replace('http://', '')
+      hostURL = hostURL.replace('https://', '')
+      if (hostURL.includes('localhost:')) {
+        origin = `http://${hostURL}`
+      } else {
+        origin = `https://${hostURL}`
+      }
+    }
+  }
+  return origin
+}
+
 export const filterProducts = (data, filter) => {
   let temp = data
   if (filter) {
